% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sri24_image_df.R
\name{sri24_image_df}
\alias{sri24_image_df}
\alias{sri24_images}
\title{SRI24 Images}
\usage{
sri24_image_df()

sri24_images()
}
\value{
A \code{data.frame} for \code{sri24_image_df} and a
character vector for \code{sri24_images}
}
\description{
SRI24 Images
}
\examples{
library(neurobase)
sri24_image_df()
imgs = sri24_images()
imgs
imgs["erly"]
img = readnii(imgs["erly"])
ortho2(img)


}
